[pod] test parallelism: per-runtime forking + runBackend API + orchestration extraction
Halves CI wall-clock for kyo-pod (~30min → ~12min) while strictly bounding
container-daemon contention to one in-flight op per daemon at a time.

Parallelism architecture:
- build.sbt testGrouping forks each suite once per runtime via KYO_POD_RUNTIME
  env var (one fork pinned to podman, one to docker), so each daemon sees a
  single-threaded test stream from a single forked JVM.
- Auto-detect which suites need the per-runtime split by instantiating each
  suite at config time and checking whether `Suite.testNames` contains the
  bracketed `[podman]` / `[docker]` markers that runBackends/runRuntimes
  register. Suites without those markers fork once with no env var pin —
  no marker trait or naming convention for humans to forget.
- ContainerRuntimeBase honours KYO_POD_RUNTIME by narrowing `available` to
  the pinned runtime, so runBackends/runRuntimes register leaves only for
  that fork's runtime.
- CI hard-caps `Tags.ForkedTestGroup` at 2 (one per daemon, no overlap).

runBackend API:
- New `runBackend` / `runBackendLong` register a single leaf using the HTTP
  backend over the auto-detected runtime socket. Use these for tests that
  exercise higher-level Container API (predefs, demos, parser-specific
  stress) where the choice of backend (HTTP vs Shell) and runtime (Podman
  vs Docker) is plumbing rather than the subject under test.
- HttpBackendStressTest: bug it reproduces is in Http1ResponseParser; shell
  backend tests nothing relevant. Migrated to runBackend (4× → 1× leaves).
- DemoTest, ContainerPredefItTest: predef and demo logic is daemon-agnostic;
  lifecycle code is exercised by ContainerItTest. Migrated to runBackend /
  runBackendLong.

Orchestration extraction:
- New ContainerOrchestrationItTest holds 10 sections + 3 loose tests that
  exercise framework code paths (kyo Async/Scope, healthcheck retry,
  multi-container orchestration helpers) where backend translation isn't
  the subject. Uses runBackend exclusively — single fork, single leaf per
  test. Cuts ContainerItTest per-fork from 11min to ~8min.

Port-conflict test fix:
- "two containers on the same host port" assumed Linux Docker semantics
  (PortConflictException raised). Docker Desktop on macOS silently accepts
  the conflicting bind and lies in inspect (NetworkSettings.Ports reports
  the requested hostPort despite no actual binding). The test now handles
  both daemon behaviours: PortConflict raised on Linux; success with the
  reported port on Docker Desktop.
